[Feature]: Fetch all card names from scryfall
