+++ app/routes/home_routes.py
@@ -1,9 +1,7 @@
-from flask import session, Blueprint, render_template, redirect, url_for, current_app
-from flask_socketio import emit
+from flask import session, Blueprint, render_template, redirect, url_for
 
-from app.caching import get_cached_card_names
 from managers.availability_manager import get_card_availability
-from managers.socket_manager.socket_manager import log_and_emit, socketio
+from managers.socket_manager.socket_manager import log_and_emit
 
 home_bp = Blueprint("home_bp", __name__)
 @home_bp.route("/")
