+++ docker-compose.yml
@@ -1,18 +1,18 @@
 services:
-  test-runner:
-    build: .
-    container_name: test-runner
-    command: python tests/test_for_explosions.py
-    depends_on:
-      - redis
-    environment:
-      - REDIS_HOST=redis
-    restart: "no"  # Prevents automatic restart on failure
-    healthcheck:  # This tells Docker Compose if the service is "healthy"
-      test: ["CMD", "test", "-f", "/app/tests_passed"]  # Only passes if the file exists
-      interval: 10s
-      retries: 3
-      start_period: 5s
+  #test-runner:
+  #  build: .
+  #  container_name: test-runner
+  #  command: python tests/test_for_explosions.py
+  #  depends_on:
+  #    - redis
+  #  environment:
+  #    - REDIS_HOST=redis
+  #  restart: "no"  # Prevents automatic restart on failure
+  #  healthcheck:  # This tells Docker Compose if the service is "healthy"
+  #    test: ["CMD", "test", "-f", "/app/tests_passed"]  # Only passes if the file exists
+  #    interval: 10s
+  #    retries: 3
+  #    start_period: 5s
 
   web-app:
     build: .
